+++ tests/test_app.py
@@ -15,7 +15,7 @@ def test_chat_prompt_json():
 def test_chat_prompt_simple():
     """Test basice prompt fuction from simple template setting."""
     app = App.from_json("tests/files/config.json")
-    app.config.inference_prompt = "System\n\n### Input:\n{question}\n\n### Context:\n{context}\n\n### Response:\n"
+    app.config.inference_prompt = "System\n\n### Input:\n{{question}}\n\n### Context:\n{{context}}\n\n### Response:\n"
     expected = "System\n\n### Input:\nQuestion\n\n### Context:\nContext\n\n### Response:\n"
     result = app.chat_prompt("Question", "Context")
     assert result == expected, "Prompt is not as expected."
@@ -25,7 +25,7 @@ def test_training_prompt_simple():
     """Test basice prompt fuction from simple template setting."""
     app = App.from_json("tests/files/config.json")
     app.config.training_prompt = (
-        "System\n\n### Input:\n{question}\n\n### Context:\n{context}\n\n### Response:\n{answer}\n\n"
+        "System\n\n### Input:\n{{question}}\n\n### Context:\n{{context}}\n\n### Response:\n{{answer}}\n\n"
     )
     expected = "System\n\n### Input:\nQuestion\n\n### Context:\nContext\n\n### Response:\nAnswer\n\n"
     result = app.training_prompt("Question", "Context", "Answer")