+++ CHANGELOG.md
@@ -23,6 +23,10 @@ With this method, the version is YEAR.MONTH.RELEASE. To increment this version, 
 
 ## [Unreleased]
 
+### Added
+
+- Added `max_training_tokens` to specify the `max_length` parameter for the tokenizer when tokenizing the training set. Defaults to the prior setting of 512.
+
 ### Fixed
 
 - load_json_arg(): Fixed bug with json string as input.