+++ CHANGELOG.md
@@ -26,12 +26,20 @@ With this method, the version is YEAR.MONTH.RELEASE. To increment this version, 
 ### Added
 
 - Added `max_training_tokens` to specify the `max_length` parameter for the tokenizer when tokenizing the training set. Defaults to the prior setting of 512.
+- Added `prompt_template_file`.
+- Added `prompt_template`.
 
 ## Changed
 
 - Logging: Switch from using logging module to loguru.
 - Template: Switch from format string to Jinja for the templates.
 - Template: Now only one template is used for both training and inference.
+- Template: Can either be set by a file path or a string.
+
+## Removed
+
+- Removed `training_prompt`. Replaced by `prompt_template`.
+- Removed `inference_prompt`. Replaced by `prompt_template`.
 
 ### Fixed